Back_end_DRF/mindbowser/requirements.txt
asgiref==3.3.1
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.5
django-cors-headers==3.6.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
drf-api-logger==0.0.9
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jwt==1.1.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
